+++ .idea/copyright/profiles_settings.xml
