--- AndroidProject/app/src/main/java/com/ssuproject/ulnamsong/androidproject/MainActivity.java
@@ -1,31 +1,0 @@
-package com.ssuproject.ulnamsong.androidproject;
-
-import android.app.AlertDialog;
-import android.content.DialogInterface;
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-import android.widget.Toast;
-
-public class MainActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
-        Toast.makeText(this, "Git Hub Test", Toast.LENGTH_SHORT).show();
-    }
-
-    public void showDialog(String name) {
-        AlertDialog.Builder alert = new AlertDialog.Builder(MainActivity.this);
-        alert.setTitle("이름 : " + name);
-        alert.setMessage(name + "가 이겼습니다.");
-        alert.setPositiveButton("OK", new DialogInterface.OnClickListener() {
-            @Override
-            public void onClick(DialogInterface dialog, int which) {
-                // Do Something
-            }
-        });
-        alert.create();
-        alert.show();
-    }
-}
