--- AndroidProject/gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,0 @@
-#Mon Nov 02 12:22:43 KST 2015
-distributionBase=GRADLE_USER_HOME
-distributionPath=wrapper/dists
-zipStoreBase=GRADLE_USER_HOME
-zipStorePath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-2.4-all.zip
