+++ app/src/main/java/com/ssuproject/ulnamsong/ssu_androidapplicationproject/MainActivity.java
@@ -1,0 +1,13 @@
+package com.ssuproject.ulnamsong.ssu_androidapplicationproject;
+
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+
+public class MainActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
+    }
+}
